+++ example/lib/avatar_group_label_section.dart
@@ -1,0 +1,56 @@
+import 'package:example/section_container.dart';
+import 'package:flowbite_ui/flowbite_ui.dart';
+import 'package:flutter/material.dart';
+
+class AvatarGroupLabelSection extends StatelessWidget {
+  const AvatarGroupLabelSection({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SectionContainer(
+      fileName: 'avatar-group-label',
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        spacing: 12.0,
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          // With subtitle
+          Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            spacing: 6.0,
+            children: FlowbiteAvatarGroupLabelSize.values
+                .map(
+                  (size) => FlowbiteAvatarGroupLabel(
+                    avatarImage: const AssetImage(
+                      'assets/images/avatar-jese-leos.png',
+                      package: 'flowbite_ui',
+                    ),
+                    title: 'Jese Leos',
+                    subtitle: 'name@flowbite.com',
+                    size: size,
+                  ),
+                )
+                .toList(),
+          ),
+          // Without subtitle
+          Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            spacing: 6.0,
+            children: FlowbiteAvatarGroupLabelSize.values
+                .map(
+                  (size) => FlowbiteAvatarGroupLabel(
+                    avatarImage: const AssetImage(
+                      'assets/images/avatar-jese-leos.png',
+                      package: 'flowbite_ui',
+                    ),
+                    title: 'Jese Leos',
+                    size: size,
+                  ),
+                )
+                .toList(),
+          ),
+        ],
+      ),
+    );
+  }
+}